requirements.txt
aiogram==3.22.0
aiosqlite==0.21.0
python-dotenv==1.1.1
qrcode[pil]==8.2
